requirements.txt
libmozdata>=0.1.35
flask>=0.11.1
flask_sqlalchemy>=2.1
flask_cors>=3.0.2
sqlalchemy>=1.1.5
gunicorn>=19.6.0
scikit-learn>=0.18.1
nltk>=3.2.4
whatthepatch>=0.0.4
python-hglib>=2.4
requests>=2.18.1
